Code review updates implemenation

+++ src/processor.cpp
@@ -6,13 +6,13 @@
 using std::vector;
 
 //( user + nice + system + idle + iowait + irq + softirq )
-float Processor::Utilization() {
+double Processor::Utilization() {
     this->total_capacity_ = LinuxParser::Jiffies();
-    float unused = static_cast<float>(LinuxParser::IdleJiffies());
+    double unused = static_cast<double>(LinuxParser::IdleJiffies());
 
-    float total = this->total_capacity_ - this->cpu_utilization[0];
-    float idle = unused - this->cpu_utilization[1];
-    float cpuPercentage = (total - idle)/total;
+    double total = this->total_capacity_ - this->cpu_utilization[0];
+    double idle = unused - this->cpu_utilization[1];
+    double cpuPercentage = (total - idle)/total;
 
     this->cpu_utilization[0] = this->total_capacity_;
     this->cpu_utilization[1] = unused;
